phantm/backend/app/middleware/auth.py
"""
Auth dependency for FastAPI routes.

Extracts and validates the JWT from the Authorization header.
Supabase Auth issues JWTs -- we verify them using the JWT secret
from Supabase project settings.
"""

from uuid import UUID

from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
from jose import JWTError, jwt

from app.config import Settings, get_settings

# Extracts "Bearer <token>" from the Authorization header
_bearer_scheme = HTTPBearer()


async def get_current_user_id(
    credentials: HTTPAuthorizationCredentials = Depends(_bearer_scheme),
    settings: Settings = Depends(get_settings),
) -> UUID:
    """Validates JWT and returns the authenticated user's UUID.

    Use as a FastAPI dependency on any protected route:
        @router.get("/something")
        async def handler(user_id: UUID = Depends(get_current_user_id)):
    """
    token = credentials.credentials

    if not settings.jwt_secret:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="JWT_SECRET not configured on server",
        )

    try:
        payload = jwt.decode(
            token,
            settings.jwt_secret,
            algorithms=["HS256"],
            audience="authenticated",
        )
    except JWTError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid or expired token",
            headers={"WWW-Authenticate": "Bearer"},
        )

    # Supabase stores the user ID in the "sub" claim
    user_id_str = payload.get("sub")
    if not user_id_str:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Token missing user identity",
        )

    return UUID(user_id_str)
