phantm/backend/requirements.txt
# LiveSelf Backend — API Server Requirements
# Runs on Railway.app or any Python 3.11 host

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
supabase==2.12.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4

# File Storage (Cloudflare R2 — S3 compatible)
boto3==1.35.86

# Cache
redis==5.2.1

# HTTP client (for calling engine service)
httpx==0.28.1

# Payments (Phase 3)
stripe==11.4.1

# Email (Phase 3)
resend==2.6.0

# Validation
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.18

# Monitoring
sentry-sdk[fastapi]==2.19.2

# Dev/Test
pytest==8.3.4
pytest-asyncio==0.24.0
